refactor/admin-role-separation - Admincontroller.grantAdminRole()
